requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-Login==0.6.3
Flask-WTF==1.2.2
greenlet==3.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
psycopg2-binary==2.9.9
python-dotenv==1.1.0
pytz==2024.2
SQLAlchemy==2.0.29
sqlalchemy-serializer==1.4.22
typing_extensions==4.13.2
Werkzeug==3.1.3
WTForms==3.2.1
